+++ turbine/plugins/borderradius.php
@@ -26,14 +26,16 @@
  * @return void
  */
 function borderradius(&$parsed){
-	global $cssp;
+	global $cssp,$browser;
+	$htc_path = rtrim(dirname($_SERVER['SCRIPT_NAME']),'/').'/plugins/borderradius/chromacorners.htc';
 	foreach($parsed as $block => $css){
 		foreach($parsed[$block] as $selector => $styles){
 			foreach($styles as $property => $values){
 				if(preg_match('/border(?:-(top|right|bottom|left)(?:-(right|left))*)*-radius/', $property, $matches)){
 					// Create the new rules and insert them
 					$borderradius_rules = borderradius_glue_rules($matches, $values);
-					$cssp->insert_properties($borderradius_rules, $block, $selector, null, $property);
+					$borderradius_rules['behavior'][] = 'url('.$htc_path.')';
+					$cssp->insert_properties($borderradius_rules, $block, $selector, $property, null);
 					// Comment the newly inserted properties
 					foreach($borderradius_rules as $border_property => $border_value){
 						CSSP::comment($parsed[$block][$selector], $border_property, 'Added by border radius plugin');
@@ -105,7 +107,7 @@ function borderradius_glue_rules($parts, $value){
 /**
  * Register the plugin
  */
-$cssp->register_plugin('before_glue', 0, 'borderradius');
+$cssp->register_plugin('borderradius', 'borderradius', 'before_glue', 0);
 
 
 ?>