+++ turbine/plugins/bugfixes.php
@@ -15,6 +15,7 @@
  * Usage: Nobrainer, just switch it on
  * Example: -
  * Status:  Stable
+ * Version: 1.0
  * Version: 1.2
  * 
  * @param mixed &$parsed
@@ -26,6 +27,8 @@ function bugfixes(&$parsed){
 
 	// IE6: Image margin bottom bug
 	$changed['img']['vertical-align'][] = 'bottom';
+	// IE: remove scrollbars from textareas
+	$changed['textarea']['overflow'][] = 'auto';
 
 	// IE6: Background image flickers on hover
 	$changed['html']['filter'][] = 'expression(document.execCommand("BackgroundImageCache",false,true))';
@@ -35,6 +38,21 @@ function bugfixes(&$parsed){
 	$changed['img']['behavior'][] = 'url("'.$htc_path.'")';
 
 	// IE6 and 7: resample images bicubic instead of using nearest neighbor method
+	if($browser->browser == 'ie' && floatval($browser->browser_version) < 7){
+		// IE6: Image margin bottom bug
+		$changed['img']['vertical-align'][] = 'bottom';
+	
+		// IE6: Background image flickers on hover
+		$changed['html']['filter'][] = 'expression(document.execCommand("BackgroundImageCache",false,true))';
+	
+		// IE6: Fix transparent PNGs, see http://www.twinhelix.com/css/iepngfix/
+		$htc_path = rtrim(dirname($_SERVER['SCRIPT_NAME']),'/').'/plugins/bugfixes/iepngfix.htc';
+		$changed['img']['behavior'][] = 'url("'.$htc_path.'")';
+		
+		// IE6: Input align, see http://tjkdesign.com/ez-css/css/base.css
+		$changed['img']['vertical-align'][] = 'text-bottom';
+	}	
+	// IE 7: resample images bicubic instead of using nearest neighbor method
 	$changed['img']['-ms-interpolation-mode'][] = 'bicubic';
 
 	// IE6 and 7: Enable full styleability for buttons, see http://www.sitepoint.com/forums/showthread.php?t=547059
@@ -43,6 +61,8 @@ function bugfixes(&$parsed){
 	$changed['button']['white-space'][] = 'nowrap';
 
 	// IE6 and 7: Missing :hover-property on every tag except a, see http://www.xs4all.nl/~peterned/csshover.html
+	$htc_path = rtrim(dirname($_SERVER['SCRIPT_NAME']),'/').'/plugins/bugfixes/csshover3.htc';
+	$changed['body']['behavior'][] = 'url("'.$htc_path.'")';
 	// IE8: Reenable cleartype where filters are set
 	$htc_path = rtrim(dirname($_SERVER['SCRIPT_NAME']),'/').'/plugins/bugfixes/';
 	$changed['body']['behavior'][] = 'url("'.$htc_path.'csshover3.htc") url("'.$htc_path.'cleartypefix.htc")';
@@ -50,6 +70,12 @@ function bugfixes(&$parsed){
 	// Firefox: Ghost margin around buttons, see http://www.sitepoint.com/forums/showthread.php?t=547059
 	$changed['button::-moz-focus-inner']['padding'][] = '0';
 	$changed['button::-moz-focus-inner']['border'][] = 'none';
+
+	// Webkit: better antialiasing, see http://maxvoltar.com/archive/-webkit-font-smoothing
+	$changed['html']['-webkit-font-smoothing'][] = 'antialiased';
+	
+	// Webkit: better kerning, see http://www.aestheticallyloyal.com/public/optimize-legibility/
+	$changed['html']['text-rendering'][] = 'optimizeLegibility';
 
 	// Add comments for the global fixes
 	foreach($changed as $selector => $styles){
@@ -61,6 +87,7 @@ function bugfixes(&$parsed){
 	// Insert the global bugfixes
 	$cssp->insert($changed, 'global');
 
+	// Apply per-element-bugfixes
 	// Apply per-block-bugfixes
 	foreach($cssp->parsed as $block => $css){
 
@@ -97,6 +124,7 @@ function bugfixes(&$parsed){
 					CSSP::comment($cssp->parsed[$block][$selector], 'position', 'Added by bugfix plugin');
 				}
 			}
+
 		}
 	}
 
@@ -106,7 +134,7 @@ function bugfixes(&$parsed){
 /**
  * Register the plugin
  */
-$cssp->register_plugin('before_compile', 0, 'bugfixes');
+$cssp->register_plugin('bugfixes', 'bugfixes', 'before_compile', 0);
 
 
 ?>